Restore OrderedMidSpline to non-lifecycle class and add OrderedMidSplineLifecycle which inherits from OrderedMidSpline.
