+++ src/navigation/planners/setup.py
@@ -20,6 +20,7 @@ setup(
     entry_points={
         "console_scripts": [
             "ordered_mid_spline = planners.node_ordered_mid_spline:main",
+            "ordered_mid_spline_lifecycle = planners.node_ordered_mid_spline_lifecycle:main",
         ],
     },
 )
